requirement.txt
python==3.10.13

matplotlib==3.8.0
pyg==2.4.0
pytorch==1.12.1
umap-learn==0.5.4